requirements.txt
fastapi
uvicorn
pydantic
python-multipart
ollama
python-docx
pymupdf
requests
streamlit